float, settings start, layout

+++ Robert/Ticket.java
@@ -1,9 +1,9 @@
 public class Ticket
 {
-    public int ticketPreis;
-    public String ticketName;
+    public float ticketPreis;
+    public String ticketName;   
 
-    public Ticket(int _ticketPreis, String _ticketName)
+    public Ticket(float _ticketPreis, String _ticketName)
     {
         ticketPreis = _ticketPreis;
         ticketName = _ticketName;
